Add a basic shader to display a grid
The shader displays a grid on the xz plane.
