+++ Assets/_Framework/Materials.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 9c812be7046794841bd3e07d06be09b3
+folderAsset: yes
+timeCreated: 1473299161
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
