+++ Assets/_Framework/Shaders.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 6d378810b0b3f4f57ae17894ea434f93
+folderAsset: yes
+timeCreated: 1473298475
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
